backend/requirements.txt
absl-py==2.1.0
aiohappyeyeballs==2.5.0
aiohttp==3.11.13
aiosignal==1.3.2
alembic==1.15.1
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.8.0
asteroid-filterbanks==0.4.0
astroid==3.3.9
async-timeout==4.0.3
asyncpg==0.30.0
attrs==25.1.0
audioread==3.0.1
bcrypt==4.0.1
cachetools==5.5.2
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
colorlog==6.9.0
contourpy==1.3.1
coverage==7.7.1
cycler==0.12.1
decorator==5.2.1
dill==0.3.9
distro==1.9.0
dnspython==2.7.0
docopt==0.6.2
dotenv==0.9.9
ecdsa==0.19.0
einops==0.8.1
email_validator==2.2.0
exceptiongroup==1.2.2
fastapi==0.115.11
filelock==3.17.0
flake8==7.1.2
fonttools==4.56.0
frozenlist==1.5.0
fsspec==2025.3.0
greenlet==3.1.1
grpcio==1.70.0
h11==0.14.0
httpcore==1.0.7
httptools==0.6.4
httpx==0.28.1
huggingface-hub==0.29.2
HyperPyYAML==1.2.2
idna==3.10
iniconfig==2.1.0
isort==6.0.1
Jinja2==3.1.6
jiter==0.8.2
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
julius==0.2.7
junit-xml-2==1.9
kiwisolver==1.4.8
langchain==0.3.20
langchain-anthropic
langchain-core==0.3.43
langchain-openai==0.3.8
langchain-text-splitters==0.3.6
langsmith==0.3.13
lazy_loader==0.4
librosa==0.10.2.post1
lightning==2.5.0.post0
lightning-utilities==0.14.0
llvmlite==0.44.0
Mako==1.3.9
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.1
mccabe==0.7.0
mdurl==0.1.2
more-itertools==10.6.0
mpmath==1.3.0
msgpack==1.1.0
multidict==6.1.0
networkx==3.4.2
numba==0.61.0
numpy==1.26.4
omegaconf==2.3.0
openai==1.65.5
openai-whisper @ git+https://github.com/openai/whisper.git@517a43ecd132a2089d85f4ebc044728a71d49f6e
optuna==4.2.1
orjson==3.10.15
outcome==1.3.0.post0
packaging==24.2
pandas==2.2.3
passlib==1.7.4
pescador==3.0.0
pillow==11.1.0
platformdirs==4.3.6
pluggy==1.5.0
pooch==1.8.2
primePy==1.3
propcache==0.3.0
protobuf==3.20.3
psycopg2-binary==2.9.10
pyannote.audio @ git+https://github.com/pyannote/pyannote-audio.git@240a7f3ef60bc613169df860b536b10e338dbf3c
pyannote.core==5.0.0
pyannote.database==5.1.3
pyannote.metrics==3.2.1
pyannote.pipeline==3.0.1
pyasn1==0.4.8
pycodestyle==2.12.1
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
pydub==0.25.1
pyflakes==3.2.0
Pygments==2.19.1
pylint==3.3.6
pylint-fail-under==0.3.0
pylint_junit==0.3.5
pyparsing==3.2.1
pytest==8.3.5
pytest-asyncio==0.26.0
pytest-cov==6.0.0
pytest-env
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-jose==3.4.0
python-multipart==0.0.20
pytorch-lightning==2.5.0.post0
pytorch-metric-learning==2.8.1
pytz==2025.1
PyYAML==6.0.2
pyzmq==26.2.1
redis==5.2.1
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
rich==13.9.4
rsa==4.9
ruamel.yaml==0.18.10
ruamel.yaml.clib==0.2.12
scikit-learn==1.6.1
scipy==1.15.2
semver==3.0.4
sentencepiece==0.2.0
shellingham==1.5.4
six>=1.12,<2
sniffio==1.3.1
sortedcollections==2.1.0
sortedcontainers==2.4.0
soundfile==0.13.1
soxr==0.5.0.post1
speechbrain==1.0.2
SQLAlchemy==2.0.38
starlette==0.46.1
starlette-context==0.4.0
sympy==1.13.3
tabulate==0.9.0
tenacity==9.0.0
tensorboard==2.19.0
tensorboard-data-server==0.7.2
tensorboardX==2.6.2.2
threadpoolctl==3.5.0
tiktoken==0.9.0
tomli==2.2.1
tomlkit==0.13.2
torch==2.2.2
torch-audiomentations==0.12.0
torch_pitch_shift==1.2.5
torchaudio==2.2.2
torchmetrics==1.6.2
tqdm==4.67.1
trio==0.29.0
typer==0.15.2
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
uvicorn==0.34.0
uvloop==0.21.0
watchfiles==1.0.4
websockets==15.0.1
Werkzeug==3.1.3
whisper==1.1.10
yarl==1.18.3
zstandard==0.23.0
langgraph
langchain_community
langchain_neo4j
langchain_chroma
langchain-huggingface
langchain-ollama
sentence-transformers
lightrag-hku
pipmaster
graspologic
pydantic-settings
uuid6
# pdfminer.six
textract==1.6.5
cryptography==44.0.3
wcwidth==0.2.13
celery==5.3.6
flower==2.0.1
fastapi-cache2==0.2.2
boto3>=1.34.0
croniter>=6.0.0
loguru>=0.7.3
opik